Update and rename CoreTest.php to TempFolderTest.php
